Make Config a trait
