+++ src/display_servers/mod.rs
@@ -5,14 +5,17 @@ use crate::models::Screen;
 use crate::models::Window;
 use crate::models::Workspace;
 use crate::DisplayEvent;
+#[cfg(test)]
 mod mock_display_server;
 pub mod xlib_display_server;
 
+#[cfg(test)]
 pub use self::mock_display_server::MockDisplayServer;
 pub use self::xlib_display_server::XlibDisplayServer;
 
-pub trait DisplayServer {
-    fn new(config: &Config) -> Self;
+pub trait DisplayServer<C: Config> {
+    fn new(config: C) -> Self;
+
     fn get_next_events(&mut self) -> Vec<DisplayEvent>;
 
     fn update_theme_settings(&mut self, _settings: ThemeSetting) {}
