+++ src/display_servers/xlib_display_server/mod.rs
@@ -24,28 +24,33 @@ mod xcursor;
 
 static SETUP: Once = Once::new();
 
-pub struct XlibDisplayServer {
+pub struct XlibDisplayServer<C: Config> {
     xw: XWrap,
     root: xlib::Window,
-    config: Config,
+    config: C,
     theme: ThemeSetting,
 }
 
-impl DisplayServer for XlibDisplayServer {
-    fn new(config: &Config) -> XlibDisplayServer {
-        let wrap = XWrap::new();
+impl<C> DisplayServer<C> for XlibDisplayServer<C>
+where
+    C: Config,
+{
+    fn new(config: C) -> Self {
+        let theme = ThemeSetting::default();
+        let mut wrap = XWrap::new();
+
+        wrap.focus_behaviour = config.focus_behaviour();
+        wrap.mouse_key_mask = utils::xkeysym_lookup::into_mod(config.mousekey());
+        wrap.init(&config, &theme); //setup events masks
+
         let root = wrap.get_default_root();
-        let mut me = XlibDisplayServer {
+
+        Self {
             xw: wrap,
             root,
-            theme: ThemeSetting::default(),
-            config: config.clone(),
-        };
-
-        me.xw.focus_behaviour = config.focus_behaviour.clone();
-        me.xw.mouse_key_mask = utils::xkeysym_lookup::into_mod(&config.mousekey);
-        me.xw.init(config, &me.theme); //setup events masks
-        me
+            theme,
+            config,
+        }
     }
 
     fn update_theme_settings(&mut self, settings: ThemeSetting) {
@@ -190,11 +195,14 @@ impl DisplayServer for XlibDisplayServer {
     }
 }
 
-impl XlibDisplayServer {
+impl<C> XlibDisplayServer<C>
+where
+    C: Config,
+{
     /// Return a vec of events for setting up state of WM.
     fn initial_events(&self) -> Vec<DisplayEvent> {
         let mut events = vec![];
-        if let Some(workspaces) = &self.config.workspaces {
+        if let Some(workspaces) = self.config.workspaces() {
             if workspaces.is_empty() {
                 // tell manager about existing screens
                 self.xw.get_screens().into_iter().for_each(|screen| {
