+++ src/display_servers/xlib_display_server/xwrap.rs
@@ -1436,7 +1436,7 @@ impl XWrap {
     }
 
     // TODO: split into smaller functions
-    pub fn init(&mut self, config: &Config, theme: &ThemeSetting) {
+    pub fn init(&mut self, config: &impl Config, theme: &ThemeSetting) {
         let root_event_mask: c_long = xlib::SubstructureRedirectMask
             | xlib::SubstructureNotifyMask
             | xlib::ButtonPressMask
